backend/frontend_communication_server/requirements.txt
flask
flask_cors
flask_sslify
waitress
psycopg2-binary
